From Thibault Genessay,
"Under Windows, the current implementation of the operator bool() of
the HandleHolder class can return true even if the handle is not
valid. This happens when Win32 calls return NULL instead of
INVALID_HANDLE_VALUE as expected by the operator bool().

The modified file makes the OpenThreads::Thread::start() method return
non-zero in this case, and should also ensure that other calls to
Win32 functions returning handles are correctly checked.
"


git-svn-id: http://www.openscenegraph.org/svn/osg/OpenThreads/trunk@4675 16af8721-9629-0410-8352-f15c8da7e697