Added OpenThreads::GetNumberOfProcessors().
Added better support for Linux processor affinity.  Note, there happens to be to
API's out in the wild for sched_setaffinity, and two param one, and a three param one.  To
arbitrate between the two ) have added a check against the define COMPILE_USING_TWO_PARAM_sched_setaffinity
in in the source.  This can be selected on the command line via:

 make COMPILE_USING_TWO_PARAM_sched_setaffinity=yes


git-svn-id: http://www.openscenegraph.org/svn/osg/OpenThreads/trunk@5417 16af8721-9629-0410-8352-f15c8da7e697